Organise scalars and collections better
